put the ring mapping the ecalDetId.py to avoid circular includes
