+++ EcalTools/macro/EndcapAsymmetry.py
@@ -1,45 +1,7 @@
 #! /usr/bin/env python
 import os
 from math import *
-from TagValidation import *
-
-
-class etaRingMapping:
-    def __init__(self):
-        Endc_x_y_ring=("%s/src/EcalReconstruction/data/Endc_x_y_ring.txt" % os.environ['CMSSW_BASE'])
-        self.eeringmap = {}
-        self.eeringcrystals = {}
-        for line in open(Endc_x_y_ring,'r'):
-            item = line.split()
-            key = (int(item[0]),int(item[1]))
-            # map x,y -> ring
-            thering = int(item[3])
-            self.eeringmap[key] = thering
-            # map ring -> vector of crystals
-            if not thering in self.eeringcrystals:
-                self.eeringcrystals[thering] = [(int(item[0]),int(item[1]))]
-            else:
-                self.eeringcrystals[thering].append((int(item[0]),int(item[1])))
-
-        self.ebringmap = {}
-        for iz in range(-1,2,2):
-            for ieta in range(1,86):
-                for iphi in range(1,361):
-                    key = (ieta,iphi,iz)
-                    self.ebringmap[key] = iz*ieta
-                    
-    def getRing(self,doEB,x,y,z):
-        if doEB: return self.ebringmap[(x,y,z)]
-        else: return self.eeringmap[(x,y)] if (x,y) in self.eeringmap else -1
-
-    def getNCrystals(self,doEB,ring):
-        if doEB: return 360
-        else: 
-            if ring==-1: return 0
-            if ring not in self.eeringcrystals: 
-                print "WARNING! Ring ",ring," not valid. Returning 0 crystals."
-                return 0
-            else: return len(self.eeringcrystals[ring])
+from ecalDetId import EcalDetId,etaRingMapping
 
 def asymmetryByRing(tv,data,doPlot=True):
